ml-service/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Machine Learning & NLP (excluding torch - install separately)
sentence-transformers
transformers
numpy
scikit-learn

# Vector Database
faiss-cpu

# PDF Processing
PyPDF2
pypdf

# HTTP Requests
requests
httpx

# Data Processing
pandas

# Logging
colorlog

# Testing (optional)
pytest
pytest-asyncio